DiffEq package extension
